let category = 'accounting'
That's what I would rather classify it as.

'hr' doesn't make a lot of sense because the menus are in accounting and
reserved for adviser, and what it really imports is accounting lines. There was
a bit of confusion as the original need for the module is to import
salary-related lines from Crésus Salaires (which are more frequently fed to
Crésus Comptabilité but it's not meant to be exclusive) in order to pay those
salaries, but the module can also be used to import lines from Crésus
Comptabilité in order to completely migrate, which has nothing to do with HR.